// src/enums.ts
export enum statusEtapa {
    pendente = "Pendente",
    andamento = "Andamento",
    concluido = "Concluido"
  }

  export enum tipoAeronave {
    comercial = "Comercial",
    militar = "Militar"
  }

  export enum tipoPeca {
    nacional = "Nacional",
    importada = "Importada"
  }

  export enum statusPeca {
    em_producao = "Em producao",
    em_transporte = "Em transporte",
    pronta = "Pronta"
  }

  export enum nivelPermissao {
    administrador = "Administrador",
    engenheiro = "Engenheiro",
    operador = "Operador"
  }

  export enum tipoTeste {
    eletrico = "Eletrico",
    hidraulico = "Hidraulico",
    aerodinamico = "Aerodinamico"
  }

  export enum resultadoTeste {
    aprovado = "Aprovado",
    reprovado = "Reprovado"
  } 